--- .github/actions/trun/deps/actions/core.ts
@@ -1,1 +1,0 @@
-export * from "npm:@actions/core"
